# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Deep Analysis Financial Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export type ScreenedStockRiskProfile =
  (typeof ScreenedStockRiskProfile)[keyof typeof ScreenedStockRiskProfile];

export const ScreenedStockRiskProfile = {
  high_growth: "high_growth",
  balanced: "balanced",
  value_stability: "value_stability",
} as const;

export interface ScreenedStock {
  symbol: string;
  name: string;
  sector: string;
  price: number;
  change: number;
  changePercent: number;
  revenueGrowth: number;
  peRatio: number;
  marketCap: number;
  institutionalOwnership: number;
  macroKeywords: string[];
  /** 0-100 geopolitical exposure score */
  geopoliticalScore: number;
  /** 0-100 LITE-style breakout score */
  breakoutScore: number;
  riskProfile: ScreenedStockRiskProfile;
  isSoldOut: boolean;
}

export interface StockQuote {
  symbol: string;
  name: string;
  price: number;
  open: number;
  high: number;
  low: number;
  volume: number;
  peRatio: number;
  eps: number;
  revenueGrowth: number;
  grossMargin: number;
  institutionalOwnership: number;
  fiftyTwoWeekHigh: number;
  fiftyTwoWeekLow: number;
  marketCap: number;
  dividendYield: number;
}

export type NewsArticleSentiment =
  (typeof NewsArticleSentiment)[keyof typeof NewsArticleSentiment];

export const NewsArticleSentiment = {
  Bullish: "Bullish",
  Bearish: "Bearish",
  Neutral: "Neutral",
  "Somewhat-Bullish": "Somewhat-Bullish",
  "Somewhat-Bearish": "Somewhat-Bearish",
} as const;

export interface NewsArticle {
  title: string;
  summary: string;
  url: string;
  source: string;
  publishedAt: string;
  sentiment: NewsArticleSentiment;
  sentimentScore: number;
  relevanceScore: number;
}

export type MarketSummaryTopSectorsItem = {
  sector: string;
  count: number;
  avgScore: number;
};

export interface MarketSummary {
  totalScreened: number;
  highGrowthCount: number;
  newDiscoveries: number;
  avgBreakoutScore: number;
  topSectors: MarketSummaryTopSectorsItem[];
  topMacroTheme: string;
}

export interface DeepDiveRequest {
  symbol: string;
  name: string;
  peRatio?: number;
  revenueGrowth?: number;
  marketCap?: number;
  institutionalOwnership?: number;
  headlines?: string[];
  geopoliticalContext?: string;
}

export type DeepDiveReportVerdict =
  (typeof DeepDiveReportVerdict)[keyof typeof DeepDiveReportVerdict];

export const DeepDiveReportVerdict = {
  Accumulate: "Accumulate",
  Hold: "Hold",
  Avoid: "Avoid",
} as const;

export interface DeepDiveReport {
  symbol: string;
  bullCase: string;
  bearCase: string;
  verdict: DeepDiveReportVerdict;
  verdictReasoning: string;
  /** 0-100 confidence in the verdict */
  confidenceScore: number;
  keyRisks: string[];
  catalysts: string[];
  generatedAt: string;
}

export type HoldingGeopoliticalRisk =
  (typeof HoldingGeopoliticalRisk)[keyof typeof HoldingGeopoliticalRisk];

export const HoldingGeopoliticalRisk = {
  Low: "Low",
  Medium: "Medium",
  High: "High",
  Critical: "Critical",
} as const;

export interface Holding {
  id: number;
  symbol: string;
  name: string;
  shares: number;
  avgCost: number;
  currentPrice: number;
  totalValue: number;
  gainLoss: number;
  gainLossPercent: number;
  geopoliticalRisk: HoldingGeopoliticalRisk;
  notes?: string;
  addedAt: string;
}

export interface CreateHoldingBody {
  symbol: string;
  name: string;
  shares: number;
  avgCost: number;
  notes?: string;
}

export interface UpdateHoldingBody {
  shares?: number;
  avgCost?: number;
  notes?: string;
}

export type HoldingsAnalysisRequestHoldingsItem = {
  symbol: string;
  name: string;
  shares: number;
};

export interface HoldingsAnalysisRequest {
  holdings: HoldingsAnalysisRequestHoldingsItem[];
}

export type HoldingsAnalysisReportOverallRisk =
  (typeof HoldingsAnalysisReportOverallRisk)[keyof typeof HoldingsAnalysisReportOverallRisk];

export const HoldingsAnalysisReportOverallRisk = {
  Low: "Low",
  Medium: "Medium",
  High: "High",
  Critical: "Critical",
} as const;

export type HoldingsAnalysisReportStockAnalysesItemGeopoliticalRisk =
  (typeof HoldingsAnalysisReportStockAnalysesItemGeopoliticalRisk)[keyof typeof HoldingsAnalysisReportStockAnalysesItemGeopoliticalRisk];

export const HoldingsAnalysisReportStockAnalysesItemGeopoliticalRisk = {
  Low: "Low",
  Medium: "Medium",
  High: "High",
  Critical: "Critical",
} as const;

export type HoldingsAnalysisReportStockAnalysesItem = {
  symbol: string;
  geopoliticalRisk: HoldingsAnalysisReportStockAnalysesItemGeopoliticalRisk;
  riskFactors: string[];
  recommendation: string;
};

export interface HoldingsAnalysisReport {
  overallRisk: HoldingsAnalysisReportOverallRisk;
  summary: string;
  stockAnalyses: HoldingsAnalysisReportStockAnalysesItem[];
  macroThreats: string[];
  opportunities: string[];
  generatedAt: string;
}

export type MacroEventCategory =
  (typeof MacroEventCategory)[keyof typeof MacroEventCategory];

export const MacroEventCategory = {
  Geopolitical: "Geopolitical",
  Trade: "Trade",
  Energy: "Energy",
  Technology: "Technology",
  Monetary: "Monetary",
  Regulatory: "Regulatory",
} as const;

export type MacroEventDirection =
  (typeof MacroEventDirection)[keyof typeof MacroEventDirection];

export const MacroEventDirection = {
  Bullish: "Bullish",
  Bearish: "Bearish",
  Neutral: "Neutral",
} as const;

export interface MacroEvent {
  id: string;
  title: string;
  category: MacroEventCategory;
  region: string;
  /** 0-100 market impact score */
  impactScore: number;
  direction: MacroEventDirection;
  affectedSectors: string[];
  description: string;
  updatedAt: string;
}

export type MacroThemeMomentum =
  (typeof MacroThemeMomentum)[keyof typeof MacroThemeMomentum];

export const MacroThemeMomentum = {
  Accelerating: "Accelerating",
  Stable: "Stable",
  Decelerating: "Decelerating",
} as const;

export interface MacroTheme {
  id: string;
  name: string;
  description: string;
  momentum: MacroThemeMomentum;
  relatedKeywords: string[];
  topStocks: string[];
}

export type ScreenStocksParams = {
  /**
   * Risk/return profile filter
   */
  riskProfile?: ScreenStocksRiskProfile;
  /**
   * Sector filter
   */
  sector?: string;
};

export type ScreenStocksRiskProfile =
  (typeof ScreenStocksRiskProfile)[keyof typeof ScreenStocksRiskProfile];

export const ScreenStocksRiskProfile = {
  high_growth: "high_growth",
  balanced: "balanced",
  value_stability: "value_stability",
} as const;

export type GetStockNewsParams = {
  limit?: number;
};