# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Deep Analysis Financial Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analysis/deep-dive:
    post:
      operationId: deepDiveAnalysis
      tags: [analysis]
      summary: Run Gemini Deep Dive analysis on a stock
      description: Sends stock data and news to Gemini Pro for bull/bear/verdict analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeepDiveRequest"
      responses:
        "200":
          description: Gemini analysis report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeepDiveReport"
  /analysis/holdings:
    post:
      operationId: analyzeHoldings
      tags: [analysis]
      summary: Analyze portfolio holdings against macro landscape
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HoldingsAnalysisRequest"
      responses:
        "200":
          description: Holdings analysis with geopolitical risk flags
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HoldingsAnalysisReport"
  /holdings:
    get:
      operationId: getHoldings
      tags: [holdings]
      summary: Get all saved portfolio holdings
      responses:
        "200":
          description: List of holdings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Holding"
    post:
      operationId: createHolding
      tags: [holdings]
      summary: Add a holding to portfolio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHoldingBody"
      responses:
        "201":
          description: Created holding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Holding"
  /holdings/{id}:
    put:
      operationId: updateHolding
      tags: [holdings]
      summary: Update a holding
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHoldingBody"
      responses:
        "200":
          description: Updated holding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Holding"
    delete:
      operationId: deleteHolding
      tags: [holdings]
      summary: Remove a holding
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    DeepDiveRequest:
      type: object
      properties:
        symbol:
          type: string
        name:
          type: string
        peRatio:
          type: number
        revenueGrowth:
          type: number
        marketCap:
          type: number
        institutionalOwnership:
          type: number
        headlines:
          type: array
          items:
            type: string
        geopoliticalContext:
          type: string
      required:
        - symbol
        - name
    DeepDiveReport:
      type: object
      properties:
        symbol:
          type: string
        bullCase:
          type: string
        bearCase:
          type: string
        verdict:
          type: string
          enum: [Accumulate, Hold, Avoid]
        verdictReasoning:
          type: string
        confidenceScore:
          type: number
          description: 0-100 confidence in the verdict
        keyRisks:
          type: array
          items:
            type: string
        catalysts:
          type: array
          items:
            type: string
        generatedAt:
          type: string
          format: date-time
      required:
        - symbol
        - bullCase
        - bearCase
        - verdict
        - verdictReasoning
        - confidenceScore
        - keyRisks
        - catalysts
        - generatedAt
    Holding:
      type: object
      properties:
        id:
          type: integer
        symbol:
          type: string
        name:
          type: string
        shares:
          type: number
        avgCost:
          type: number
        currentPrice:
          type: number
        totalValue:
          type: number
        gainLoss:
          type: number
        gainLossPercent:
          type: number
        geopoliticalRisk:
          type: string
          enum: [Low, Medium, High, Critical]
        notes:
          type: string
        addedAt:
          type: string
          format: date-time
      required:
        - id
        - symbol
        - name
        - shares
        - avgCost
        - currentPrice
        - totalValue
        - gainLoss
        - gainLossPercent
        - geopoliticalRisk
        - addedAt
    CreateHoldingBody:
      type: object
      properties:
        symbol:
          type: string
        name:
          type: string
        shares:
          type: number
        avgCost:
          type: number
        notes:
          type: string
      required:
        - symbol
        - name
        - shares
        - avgCost
    UpdateHoldingBody:
      type: object
      properties:
        shares:
          type: number
        avgCost:
          type: number
        notes:
          type: string
    HoldingsAnalysisRequest:
      type: object
      properties:
        holdings:
          type: array
          items:
            type: object
            properties:
              symbol:
                type: string
              name:
                type: string
              shares:
                type: number
            required:
              - symbol
              - name
              - shares
      required:
        - holdings
    HoldingsAnalysisReport:
      type: object
      properties:
        overallRisk:
          type: string
          enum: [Low, Medium, High, Critical]
        summary:
          type: string
        stockAnalyses:
          type: array
          items:
            type: object
            properties:
              symbol:
                type: string
              geopoliticalRisk:
                type: string
                enum: [Low, Medium, High, Critical]
              riskFactors:
                type: array
                items:
                  type: string
              recommendation:
                type: string
            required:
              - symbol
              - geopoliticalRisk
              - riskFactors
              - recommendation
        macroThreats:
          type: array
          items:
            type: string
        opportunities:
          type: array
          items:
            type: string
        generatedAt:
          type: string
          format: date-time
      required:
        - overallRisk
        - summary
        - stockAnalyses
        - macroThreats
        - opportunities
        - generatedAt
...