# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Deep Analysis Financial Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns stocks matching the LITE-style breakout criteria
 * @summary Screen high-growth infrastructure stocks
 */
export const ScreenStocksQueryParams = zod.object({
  riskProfile: zod
    .enum(["high_growth", "balanced", "value_stability"])
    .optional()
    .describe("Risk\/return profile filter"),
  sector: zod.coerce.string().optional().describe("Sector filter"),
});

export const ScreenStocksResponseItem = zod.object({
  symbol: zod.string(),
  name: zod.string(),
  sector: zod.string(),
  price: zod.number(),
  change: zod.number(),
  changePercent: zod.number(),
  revenueGrowth: zod.number(),
  peRatio: zod.number(),
  marketCap: zod.number(),
  institutionalOwnership: zod.number(),
  macroKeywords: zod.array(zod.string()),
  geopoliticalScore: zod.number().describe("0-100 geopolitical exposure score"),
  breakoutScore: zod.number().describe("0-100 LITE-style breakout score"),
  riskProfile: zod.enum(["high_growth", "balanced", "value_stability"]),
  isSoldOut: zod.boolean(),
});
export const ScreenStocksResponse = zod.array(ScreenStocksResponseItem);

/**
 * @summary Get real-time stock quote and fundamentals
 */
export const GetStockQuoteParams = zod.object({
  symbol: zod.coerce.string(),
});

export const GetStockQuoteResponse = zod.object({
  symbol: zod.string(),
  name: zod.string(),
  price: zod.number(),
  open: zod.number(),
  high: zod.number(),
  low: zod.number(),
  volume: zod.number(),
  peRatio: zod.number(),
  eps: zod.number(),
  revenueGrowth: zod.number(),
  grossMargin: zod.number(),
  institutionalOwnership: zod.number(),
  fiftyTwoWeekHigh: zod.number(),
  fiftyTwoWeekLow: zod.number(),
  marketCap: zod.number(),
  dividendYield: zod.number(),
});

/**
 * @summary Get latest news for a stock
 */
export const GetStockNewsParams = zod.object({
  symbol: zod.coerce.string(),
});

export const getStockNewsQueryLimitDefault = 10;

export const GetStockNewsQueryParams = zod.object({
  limit: zod.coerce.number().default(getStockNewsQueryLimitDefault),
});

export const GetStockNewsResponseItem = zod.object({
  title: zod.string(),
  summary: zod.string(),
  url: zod.string(),
  source: zod.string(),
  publishedAt: zod.string(),
  sentiment: zod.enum([
    "Bullish",
    "Bearish",
    "Neutral",
    "Somewhat-Bullish",
    "Somewhat-Bearish",
  ]),
  sentimentScore: zod.number(),
  relevanceScore: zod.number(),
});
export const GetStockNewsResponse = zod.array(GetStockNewsResponseItem);

/**
 * Total screened, top movers, sector breakdowns
 * @summary Get market summary stats
 */
export const GetMarketSummaryResponse = zod.object({
  totalScreened: zod.number(),
  highGrowthCount: zod.number(),
  newDiscoveries: zod.number(),
  avgBreakoutScore: zod.number(),
  topSectors: zod.array(
    zod.object({
      sector: zod.string(),
      count: zod.number(),
      avgScore: zod.number(),
    }),
  ),
  topMacroTheme: zod.string(),
});

/**
 * Sends stock data and news to Gemini Pro for bull/bear/verdict analysis
 * @summary Run Gemini Deep Dive analysis on a stock
 */
export const DeepDiveAnalysisBody = zod.object({
  symbol: zod.string(),
  name: zod.string(),
  peRatio: zod.number().optional(),
  revenueGrowth: zod.number().optional(),
  marketCap: zod.number().optional(),
  institutionalOwnership: zod.number().optional(),
  headlines: zod.array(zod.string()).optional(),
  geopoliticalContext: zod.string().optional(),
});

export const DeepDiveAnalysisResponse = zod.object({
  symbol: zod.string(),
  bullCase: zod.string(),
  bearCase: zod.string(),
  verdict: zod.enum(["Accumulate", "Hold", "Avoid"]),
  verdictReasoning: zod.string(),
  confidenceScore: zod.number().describe("0-100 confidence in the verdict"),
  keyRisks: zod.array(zod.string()),
  catalysts: zod.array(zod.string()),
  generatedAt: zod.coerce.date(),
});

/**
 * @summary Analyze portfolio holdings against macro landscape
 */
export const AnalyzeHoldingsBody = zod.object({
  holdings: zod.array(
    zod.object({
      symbol: zod.string(),
      name: zod.string(),
      shares: zod.number(),
    }),
  ),
});

export const AnalyzeHoldingsResponse = zod.object({
  overallRisk: zod.enum(["Low", "Medium", "High", "Critical"]),
  summary: zod.string(),
  stockAnalyses: zod.array(
    zod.object({
      symbol: zod.string(),
      geopoliticalRisk: zod.enum(["Low", "Medium", "High", "Critical"]),
      riskFactors: zod.array(zod.string()),
      recommendation: zod.string(),
    }),
  ),
  macroThreats: zod.array(zod.string()),
  opportunities: zod.array(zod.string()),
  generatedAt: zod.coerce.date(),
});

/**
 * @summary Get all saved portfolio holdings
 */
export const GetHoldingsResponseItem = zod.object({
  id: zod.number(),
  symbol: zod.string(),
  name: zod.string(),
  shares: zod.number(),
  avgCost: zod.number(),
  currentPrice: zod.number(),
  totalValue: zod.number(),
  gainLoss: zod.number(),
  gainLossPercent: zod.number(),
  geopoliticalRisk: zod.enum(["Low", "Medium", "High", "Critical"]),
  notes: zod.string().optional(),
  addedAt: zod.coerce.date(),
});
export const GetHoldingsResponse = zod.array(GetHoldingsResponseItem);

/**
 * @summary Add a holding to portfolio
 */
export const CreateHoldingBody = zod.object({
  symbol: zod.string(),
  name: zod.string(),
  shares: zod.number(),
  avgCost: zod.number(),
  notes: zod.string().optional(),
});

/**
 * @summary Update a holding
 */
export const UpdateHoldingParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateHoldingBody = zod.object({
  shares: zod.number().optional(),
  avgCost: zod.number().optional(),
  notes: zod.string().optional(),
});

export const UpdateHoldingResponse = zod.object({
  id: zod.number(),
  symbol: zod.string(),
  name: zod.string(),
  shares: zod.number(),
  avgCost: zod.number(),
  currentPrice: zod.number(),
  totalValue: zod.number(),
  gainLoss: zod.number(),
  gainLossPercent: zod.number(),
  geopoliticalRisk: zod.enum(["Low", "Medium", "High", "Critical"]),
  notes: zod.string().optional(),
  addedAt: zod.coerce.date(),
});

/**
 * @summary Remove a holding
 */
export const DeleteHoldingParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * Returns current geopolitical events and their market impact scores
 * @summary Get macro geopolitical heatmap data
 */
export const GetMacroHeatmapResponseItem = zod.object({
  id: zod.string(),
  title: zod.string(),
  category: zod.enum([
    "Geopolitical",
    "Trade",
    "Energy",
    "Technology",
    "Monetary",
    "Regulatory",
  ]),
  region: zod.string(),
  impactScore: zod.number().describe("0-100 market impact score"),
  direction: zod.enum(["Bullish", "Bearish", "Neutral"]),
  affectedSectors: zod.array(zod.string()),
  description: zod.string(),
  updatedAt: zod.coerce.date(),
});
export const GetMacroHeatmapResponse = zod.array(GetMacroHeatmapResponseItem);

/**
 * @summary Get active macro themes driving the market
 */
export const GetMacroThemesResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  description: zod.string(),
  momentum: zod.enum(["Accelerating", "Stable", "Decelerating"]),
  relatedKeywords: zod.array(zod.string()),
  topStocks: zod.array(zod.string()),
});
export const GetMacroThemesResponse = zod.array(GetMacroThemesResponseItem);