# lib/api-zod/src/generated/types/createHoldingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Deep Analysis Financial Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateHoldingBody {
  symbol: string;
  name: string;
  shares: number;
  avgCost: number;
  notes?: string;
}