# lib/api-zod/src/generated/types/holding.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Deep Analysis Financial Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { HoldingGeopoliticalRisk } from "./holdingGeopoliticalRisk";

export interface Holding {
  id: number;
  symbol: string;
  name: string;
  shares: number;
  avgCost: number;
  currentPrice: number;
  totalValue: number;
  gainLoss: number;
  gainLossPercent: number;
  geopoliticalRisk: HoldingGeopoliticalRisk;
  notes?: string;
  addedAt: Date;
}