# lib/api-zod/src/generated/types/updateHoldingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Deep Analysis Financial Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateHoldingBody {
  shares?: number;
  avgCost?: number;
  notes?: string;
}